Refactored name to follow angular convention & made get, post, delete functions public
