+++ src/app/services/server-api.service.ts
@@ -10,8 +10,16 @@ export class ServerApiService {
   constructor(private http: HttpClient) { }
 
   // Base Function
-  private getFromDataBase(urlExtension: string): Observable<any> {
+  public getFromDataBase(urlExtension: string): Observable<any> {
     return this.http.get(this.baseUrl + urlExtension, this.createHeaders());
+  }
+
+  public postToDataBase(urlExtension: string, body: any): Observable<any> {
+    return this.http.post(this.baseUrl + urlExtension, body, this.createHeaders());
+  }
+
+  public deleteFromDataBase(urlExtension: string): Observable<any> {
+    return this.http.delete(this.baseUrl + urlExtension, this.createHeaders());
   }
 
   private createHeaders() {
@@ -24,6 +32,7 @@ export class ServerApiService {
   }
 
   // TODO: Add appropriate authorization that should be collected when a Audiologist or Admin Logs in.
+  // TODO: We will pull this into sepperate https interceptor at a future point.
   private generateAuthorizationToken(): string {
     return ' ';
   }
